Add server-side permission validations (#30)

+++ server/citizens/citizens.lua
@@ -19,6 +19,13 @@ if not config.citizens.synced then
     )
 
     lib.callback.register("evidences:storeCitizen", function(source, arguments)
+        if not framework.hasPermission(config.permissions.access, source) then
+            return {
+                success = false,
+                response = "laptop.notifications.no_permission.description"
+            }
+        end
+
         if arguments.identifier then
             return database.update(
                 [[
@@ -48,6 +55,13 @@ if not config.citizens.synced then
     end)
 
     lib.callback.register("evidences:deleteCitizen", function(source, arguments)
+        if not framework.hasPermission(config.permissions.access, source) then
+            return {
+                success = false,
+                response = "laptop.notifications.no_permission.description"
+            }
+        end
+
         local identifier <const> = arguments.identifier
 
         local result = database.update("DELETE FROM linked_fingerprint WHERE identifier = ?", identifier)
@@ -63,23 +77,29 @@ if not config.citizens.synced then
 end
 
 lib.callback.register("evidences:getCitizens", function(source, arguments)
+    if not framework.hasPermission(config.permissions.access, source) then
+        return {
+            success = false,
+            response = "laptop.notifications.no_permission.description"
+        }
+    end
+
     arguments.searchText = arguments.searchText or ""
-    arguments.limit = arguments.limit or 1
     arguments.offset = arguments.offset or 0
 
     if config.citizens.synced then
-        return framework.getCitizens(arguments.searchText, arguments.limit, arguments.offset)
+        return framework.getCitizens(arguments.searchText, arguments.offset)
     end
 
-    local pattern <const> = "%" .. arguments.searchText:gsub("\\", "\\\\"):gsub("%%", "\\%%"):gsub("_", "\\_") .. "%"
+    local pattern <const> = "%" .. arguments.searchText:sub(1, 25):gsub("\\", "\\\\"):gsub("%%", "\\%%"):gsub("_", "\\_") .. "%"
 
     return database.query(
         [[
             SELECT * FROM citizens
             HAVING fullName LIKE ?
-            LIMIT ? OFFSET ?
+            LIMIT 10 OFFSET ?
         ]],
-        pattern, arguments.limit, arguments.offset
+        pattern, arguments.offset
     )
 end)
 
@@ -96,6 +116,13 @@ function citizens.getCitizen(identifier)
 end
 
 lib.callback.register("evidences:getCitizen", function(source, arguments)
+    if not framework.hasPermission(config.permissions.access, source) then
+        return {
+            success = false,
+            response = "laptop.notifications.no_permission.description"
+        }
+    end
+    
     return citizens.getCitizen(arguments.identifier)
 end)
